mark the vmhelper pipes as close-on-exec.
a second vmnet helper device may be created (and the child spawned) before
the first one is shut down. in this, case the new child could inherit
the first set of pipes, keeping them open and preventing the first child
from closing (as the stdin pipe remains open).

replace global_alloc w/new to match upstream as of
4f05917494b22943ab0e7c466e4eb9f7a5896daa