konamigv.cpp: Added option to use joystick/button inputs  for weddingr.
Control panel type can be selected in service mode.  This allows playing
with joystick/button inputs, and also allows the completely unused
second and third button inputs to be tested in service mode.  It's still
possible to navigate service mode and correct the control panel setting
if the wrong type is selected.